refactor: split init generation helpers
