+++ src/init/generation.rs
@@ -1,613 +1,19 @@
-use anyhow::{Context, Result};
+use anyhow::Result;
 use std::io::{self, BufRead, Write};
 use std::path::Path;
 use std::process::Command;
 
 use super::LinterGroup;
-use super::detection::parse_tool_keys;
 
-/// Returns the renovate preset entry to inject, e.g. `github>grafana/flint#v0.9.2`.
-/// Pre-release suffixes are stripped so dev builds produce a valid tag reference.
-pub(super) fn flint_preset() -> String {
-    let ver = env!("CARGO_PKG_VERSION");
-    let ver = ver.split('-').next().unwrap_or(ver);
-    format!("github>grafana/flint#v{ver}")
-}
-
-/// Adds the flint renovate preset to the `extends` array in a renovate config file.
-/// Works for both JSON and JSON5. If an unpinned or differently-pinned flint entry
-/// already exists, it is replaced in-place rather than duplicated.
-/// Returns `true` if the file was changed.
-pub(super) fn patch_renovate_extends(path: &Path) -> Result<bool> {
-    let entry = flint_preset();
-    let content = std::fs::read_to_string(path)?;
-
-    if content.contains(&entry) {
-        return Ok(false);
-    }
-
-    // If an existing flint entry (any pin) is present, replace it in-place.
-    const FLINT_ENTRY_PREFIX: &str = "\"github>grafana/flint";
-    let new_content = if let Some(pos) = content.find(FLINT_ENTRY_PREFIX) {
-        let after_open = pos + 1; // skip leading "
-        let close = content[after_open..]
-            .find('"')
-            .context("unclosed quote in existing flint preset entry")?;
-        let end = after_open + close + 1; // position after closing "
-        format!("{}\"{}\"{}", &content[..pos], entry, &content[end..])
-    } else {
-        add_to_extends(&content, &entry)
-            .with_context(|| format!("failed to patch extends in {}", path.display()))?
-    };
-
-    std::fs::write(path, new_content)?;
-    Ok(true)
-}
-
-/// Text-based insertion of `entry` into the `extends` array.
-/// Works for both JSON (`"extends": [`) and JSON5 (`extends: [`).
-fn add_to_extends(content: &str, entry: &str) -> Result<String> {
-    let re = regex::Regex::new(r#"(?:"extends"|extends)\s*:\s*\["#).unwrap();
-
-    if let Some(m) = re.find(content) {
-        let bracket_pos = m.end() - 1; // index of '['
-        let inside_start = bracket_pos + 1;
-
-        let close_offset = content[inside_start..]
-            .find(']')
-            .context("extends array has no closing ]")?;
-        let close_pos = inside_start + close_offset;
-        let inside = &content[inside_start..close_pos];
-
-        if inside.contains('\n') {
-            // Multiline: detect indent from first non-empty line, insert at top
-            let indent = inside
-                .lines()
-                .find(|l| !l.trim().is_empty())
-                .map(|l| " ".repeat(l.len() - l.trim_start().len()))
-                .unwrap_or_else(|| "  ".to_string());
-            Ok(format!(
-                "{}\n{}\"{}\"{}{}",
-                &content[..inside_start],
-                indent,
-                entry,
-                ",",
-                &content[inside_start..]
-            ))
-        } else {
-            // Single-line (empty or not): prepend entry
-            let sep = if inside.trim().is_empty() { "" } else { ", " };
-            Ok(format!(
-                "{}\"{}\"{}{}",
-                &content[..inside_start],
-                entry,
-                sep,
-                &content[inside_start..]
-            ))
-        }
-    } else {
-        // No extends key — add after the opening {
-        let open = content
-            .find('{')
-            .context("no opening { in renovate config")?;
-        let (before, after) = content.split_at(open + 1);
-        Ok(format!(
-            "{}\n  \"extends\": [\"{}\"],{}",
-            before, entry, after
-        ))
-    }
-}
-
-/// Runs `mise use --pin <key>@<version>` in the project directory to add a tool
-/// with a pinned version (mise resolves `latest`/`lts` to a concrete version at
-/// write time). Returns `true` if the key was written to the config (checked by
-/// re-reading the file), ignoring non-zero exit codes that arise from post-write
-/// steps like shim rebuilds failing in restricted environments.
-fn pin_tool_via_mise(project_root: &Path, key: &str, version: &str) -> bool {
-    let mise_path = project_root.join("mise.toml");
-    let before = std::fs::read_to_string(&mise_path).unwrap_or_default();
-
-    let _ = Command::new("mise")
-        .args(["use", "--pin", &format!("{key}@{version}")])
-        .current_dir(project_root)
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .status();
-
-    // Success = the key is now present in the config (regardless of exit code).
-    let after = std::fs::read_to_string(&mise_path).unwrap_or_default();
-    after != before && parse_tool_keys(&after).contains(key)
-}
-
-/// True when `[tools]` contains at least one `npm:*` key but no `node` entry.
-/// The npm backend needs a Node.js runtime; without an explicit pin, mise falls
-/// back to system node — may be absent, wrong version, or drift across machines.
-pub(crate) fn needs_node_for_npm(content: &str) -> bool {
-    let Ok(doc) = content.parse::<toml_edit::DocumentMut>() else {
-        return false;
-    };
-    let Some(tools) = doc.get("tools").and_then(|t| t.as_table()) else {
-        return false;
-    };
-    let has_npm = tools.iter().any(|(k, _)| k.starts_with("npm:"));
-    let has_node = tools.contains_key("node");
-    has_npm && !has_node
-}
-
-/// Ensures a `node` entry exists in mise.toml when any `npm:*` backend tool is
-/// present. Prefers `mise use --pin node@lts` so the version resolves to a
-/// concrete release at write time; falls back to writing `node = "lts"` directly
-/// via toml_edit when mise isn't available.
-///
-/// Returns `true` if a `node` entry was added.
-pub(crate) fn ensure_node_for_npm(project_root: &Path) -> Result<bool> {
-    let mise_path = project_root.join("mise.toml");
-    let content = std::fs::read_to_string(&mise_path).unwrap_or_default();
-    if !needs_node_for_npm(&content) {
-        return Ok(false);
-    }
-    if pin_tool_via_mise(project_root, "node", "lts") {
-        return Ok(true);
-    }
-    // Fallback: write a soft pin so mise.toml still declares the prereq.
-    let mut doc: toml_edit::DocumentMut = content.parse().context("failed to parse mise.toml")?;
-    let tools = doc["tools"]
-        .as_table_mut()
-        .context("[tools] is not a table")?;
-    tools.insert("node", toml_edit::value("lts"));
-    std::fs::write(&mise_path, doc.to_string())?;
-    Ok(true)
-}
-
-/// Pins `github:grafana/flint` in mise.toml at the calling binary's version so
-/// contributors all run the same flint release. Skips when the key already
-/// exists (any pin — never overwrite the user's explicit choice). Pre-release
-/// suffixes are stripped to match [`flint_preset`].
-///
-/// Returns `true` if a flint entry was added.
-pub(crate) fn ensure_flint_self_pin(project_root: &Path, flint_rev: Option<&str>) -> Result<bool> {
-    const RELEASE_KEY: &str = "github:grafana/flint";
-    const CARGO_KEY: &str = "cargo:https://github.com/grafana/flint";
-    let mise_path = project_root.join("mise.toml");
-    let content = std::fs::read_to_string(&mise_path).unwrap_or_default();
-    let ver = env!("CARGO_PKG_VERSION");
-    let ver = ver.split('-').next().unwrap_or(ver);
-    let mut doc: toml_edit::DocumentMut = if content.is_empty() {
-        "[tools]\n".parse().unwrap()
-    } else {
-        content.parse().context("failed to parse mise.toml")?
-    };
-    if doc.get("tools").is_none() {
-        doc["tools"] = toml_edit::table();
-    }
-    let keys_to_remove = doc
-        .get("tools")
-        .and_then(|t| t.as_table())
-        .map(|tools| {
-            tools
-                .iter()
-                .filter_map(|(key, _)| {
-                    (is_flint_tool_key(key)
-                        && (flint_rev.is_some() || key != RELEASE_KEY)
-                        && (flint_rev.is_none() || key != CARGO_KEY))
-                        .then_some(key.to_string())
-                })
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-
-    let removing_flint_key = !keys_to_remove.is_empty();
-    let mut changed = false;
-    let tools = doc["tools"]
-        .as_table_mut()
-        .context("[tools] is not a table")?;
-    for key in keys_to_remove {
-        if tools.remove(&key).is_some() {
-            changed = true;
-        }
-    }
-
-    match flint_rev {
-        Some(rev) => {
-            let rev = rev.strip_prefix("rev:").unwrap_or(rev);
-            let value = format!("rev:{rev}");
-            if tools.get(CARGO_KEY).and_then(|item| item.as_str()) != Some(value.as_str()) {
-                tools.insert(CARGO_KEY, toml_edit::value(value));
-                changed = true;
-            }
-        }
-        None => {
-            if !tools.contains_key(RELEASE_KEY) {
-                if !removing_flint_key && pin_tool_via_mise(project_root, RELEASE_KEY, ver) {
-                    return Ok(true);
-                } else {
-                    tools.insert(RELEASE_KEY, toml_edit::value(ver));
-                    changed = true;
-                }
-            }
-        }
-    }
-
-    if changed {
-        std::fs::write(&mise_path, doc.to_string())?;
-    }
-    Ok(changed)
-}
-
-fn is_flint_tool_key(key: &str) -> bool {
-    key == "github:grafana/flint"
-        || key.starts_with("cargo:https://github.com/grafana/flint")
-        || key.starts_with("cargo:https://github.com/grafana/flint.git")
-}
-
-/// Replaces obsolete tool keys in mise.toml with their modern equivalents,
-/// preserving the existing version value. Returns the list of replacements made
-/// as `(old_key, new_key)` pairs. No-ops if the file doesn't exist or has no
-/// obsolete keys.
-pub fn replace_obsolete_keys(
-    project_root: &Path,
-    obsolete: &[(&str, &str)],
-) -> Result<Vec<(String, String)>> {
-    let path = project_root.join("mise.toml");
-    let content = match std::fs::read_to_string(&path) {
-        Ok(c) => c,
-        Err(e) if e.kind() == io::ErrorKind::NotFound => return Ok(vec![]),
-        Err(e) => return Err(e).with_context(|| format!("failed to read {}", path.display())),
-    };
-    let mut doc: toml_edit::DocumentMut = content.parse().context("failed to parse mise.toml")?;
-
-    let mut replaced = vec![];
-    if let Some(tools) = doc.get_mut("tools").and_then(|t| t.as_table_mut()) {
-        for &(old_key, new_key) in obsolete {
-            if let Some(value) = tools.remove(old_key) {
-                tools.insert(new_key, value);
-                replaced.push((old_key.to_string(), new_key.to_string()));
-            }
-        }
-    }
-
-    if !replaced.is_empty() {
-        std::fs::write(&path, doc.to_string()).context("failed to write mise.toml")?;
-    }
-    Ok(replaced)
-}
-
-/// Removes specific tool keys from mise.toml.
-/// Returns the list of removed keys. No-ops if the file doesn't exist or none are present.
-pub fn remove_tool_keys(project_root: &Path, keys: &[&str]) -> Result<Vec<String>> {
-    let path = project_root.join("mise.toml");
-    let content = match std::fs::read_to_string(&path) {
-        Ok(c) => c,
-        Err(e) if e.kind() == io::ErrorKind::NotFound => return Ok(vec![]),
-        Err(e) => return Err(e).with_context(|| format!("failed to read {}", path.display())),
-    };
-    let mut doc: toml_edit::DocumentMut = content.parse().context("failed to parse mise.toml")?;
-
-    let mut removed = vec![];
-    if let Some(tools) = doc.get_mut("tools").and_then(|t| t.as_table_mut()) {
-        for key in keys {
-            if tools.remove(key).is_some() {
-                removed.push((*key).to_string());
-            }
-        }
-    }
-
-    if !removed.is_empty() {
-        std::fs::write(&path, doc.to_string()).context("failed to write mise.toml")?;
-    }
-    Ok(removed)
-}
-
-pub(super) fn apply_changes(
-    path: &Path,
-    current_content: &str,
-    to_add: &[(String, Option<String>)],
-    to_remove: &[String],
-    to_upgrade: &[(String, String)],
-) -> Result<()> {
-    let project_root = path.parent().unwrap_or(path);
-
-    // Pin new tools via `mise use --pin`. For tools where mise succeeds the
-    // file is already updated; we still open the file below to handle removals,
-    // upgrades, and component additions.
-    let mut pinned_via_mise: std::collections::HashSet<String> = std::collections::HashSet::new();
-    for (key, _) in to_add {
-        if pin_tool_via_mise(project_root, key, "latest") {
-            pinned_via_mise.insert(key.clone());
-        } else {
-            eprintln!("  warning: could not pin {key} via mise — writing \"latest\"");
-        }
-    }
-
-    // Re-read the file only if mise actually modified it.
-    let current_content: String = if pinned_via_mise.is_empty() {
-        current_content.to_string()
-    } else {
-        std::fs::read_to_string(path).unwrap_or_else(|_| current_content.to_string())
-    };
-    let mut doc: toml_edit::DocumentMut = current_content
-        .parse()
-        .unwrap_or_else(|_| toml_edit::DocumentMut::new());
-
-    // Ensure [tools] table exists.
-    if !doc.contains_key("tools") {
-        doc.insert("tools", toml_edit::Item::Table(toml_edit::Table::new()));
-    }
-    let tools = doc["tools"]
-        .as_table_mut()
-        .context("[tools] is not a table")?;
-
-    for key in to_remove {
-        tools.remove(key.as_str());
-    }
-
-    for (key, components) in to_add {
-        let already_pinned = pinned_via_mise.contains(key.as_str());
-        match components {
-            Some(comps) => {
-                // If mise already wrote a plain-string version, upgrade to inline
-                // table to attach the components field.
-                let existing_version = if already_pinned {
-                    tools
-                        .get(key.as_str())
-                        .and_then(|i| i.as_value())
-                        .and_then(|v| match v {
-                            toml_edit::Value::String(s) => Some(s.value().to_string()),
-                            toml_edit::Value::InlineTable(t) => t
-                                .get("version")
-                                .and_then(|v| v.as_str())
-                                .map(str::to_string),
-                            _ => None,
-                        })
-                        .unwrap_or_else(|| "latest".to_string())
-                } else {
-                    "latest".to_string()
-                };
-                let mut tbl = toml_edit::InlineTable::new();
-                tbl.insert("version", toml_edit::Value::from(existing_version.as_str()));
-                tbl.insert("components", toml_edit::Value::from(comps.as_str()));
-                tools.insert(
-                    key.as_str(),
-                    toml_edit::Item::Value(toml_edit::Value::InlineTable(tbl)),
-                );
-            }
-            None => {
-                if !already_pinned {
-                    tools.insert(key.as_str(), toml_edit::value("latest"));
-                }
-                // Already pinned by mise — leave the entry as-is.
-            }
-        }
-    }
-
-    // Upgrade existing entries: preserve the current version, update components.
-    for (key, components) in to_upgrade {
-        let existing_version = tools
-            .get(key.as_str())
-            .and_then(|item| item.as_value())
-            .and_then(|v| match v {
-                toml_edit::Value::String(s) => Some(s.value().to_string()),
-                toml_edit::Value::InlineTable(tbl) => tbl
-                    .get("version")
-                    .and_then(|v| v.as_str())
-                    .map(str::to_string),
-                _ => None,
-            })
-            .unwrap_or_else(|| "latest".to_string());
-
-        let mut tbl = toml_edit::InlineTable::new();
-        tbl.insert("version", toml_edit::Value::from(existing_version.as_str()));
-        tbl.insert("components", toml_edit::Value::from(components.as_str()));
-        tools.insert(
-            key.as_str(),
-            toml_edit::Item::Value(toml_edit::Value::InlineTable(tbl)),
-        );
-    }
-
-    std::fs::write(path, doc.to_string())?;
-    Ok(())
-}
-
-/// Sorts `[tools]` entries and inserts the `# Linters` header when they are not
-/// already in canonical form. Returns `true` if the file was rewritten.
-fn normalize_tools_section_impl(path: &Path, verbose: bool) -> Result<bool> {
-    let content = match std::fs::read_to_string(path) {
-        Ok(c) => c,
-        Err(_) => return Ok(false),
-    };
-    let mut doc: toml_edit::DocumentMut = match content.parse() {
-        Ok(d) => d,
-        Err(_) => return Ok(false),
-    };
-    let Some(tools) = doc.get_mut("tools").and_then(|i| i.as_table_mut()) else {
-        return Ok(false);
-    };
-    sort_and_group_tools(tools, &content);
-    let new_content = doc.to_string();
-    if new_content == content {
-        return Ok(false);
-    }
-    std::fs::write(path, new_content)?;
-    if verbose {
-        println!("  normalized [tools] in {}", path.display());
-    }
-    Ok(true)
-}
-
-pub(crate) fn tools_section_needs_normalization(path: &Path) -> Result<bool> {
-    let content = match std::fs::read_to_string(path) {
-        Ok(c) => c,
-        Err(e) if e.kind() == io::ErrorKind::NotFound => return Ok(false),
-        Err(e) => return Err(e).with_context(|| format!("failed to read {}", path.display())),
-    };
-    let mut doc: toml_edit::DocumentMut = content
-        .parse()
-        .with_context(|| format!("failed to parse {}", path.display()))?;
-    let Some(tools) = doc.get_mut("tools").and_then(|i| i.as_table_mut()) else {
-        return Ok(false);
-    };
-    sort_and_group_tools(tools, &content);
-    Ok(doc.to_string() != content)
-}
-
-pub(crate) fn normalize_tools_section(path: &Path) -> Result<bool> {
-    normalize_tools_section_impl(path, true)
-}
-
-/// Sorts `[tools]` entries alphabetically and inserts a `# Linters` comment
-/// before the first linter entry. Runtime, SDK, and unrelated project tools stay
-/// above the header; known linter keys go below.
-fn sort_and_group_tools(tools: &mut toml_edit::Table, original: &str) {
-    let mut entries: Vec<(String, toml_edit::Item)> = tools
-        .iter()
-        .map(|(k, v)| (k.to_string(), v.clone()))
-        .collect();
-    if entries.is_empty() {
-        return;
-    }
-    let linter_keys = crate::registry::linter_keys();
-    let (mut linters, mut runtimes): (Vec<_>, Vec<_>) = entries
-        .drain(..)
-        .partition(|(k, _)| linter_keys.contains(k.as_str()));
-    runtimes.sort_by(|a, b| a.0.cmp(&b.0));
-    linters.sort_by(|a, b| a.0.cmp(&b.0));
-    let preserved_prefixes: std::collections::HashMap<String, String> = tools
-        .iter()
-        .filter_map(|(k, _)| {
-            let prefix = tools
-                .key(k)
-                .and_then(|key| key.leaf_decor().prefix())
-                .and_then(|prefix| {
-                    prefix
-                        .as_str()
-                        .or_else(|| prefix.span().and_then(|span| original.get(span)))
-                })?;
-            let is_header = prefix == "\n# Linters\n";
-            (!is_header && !prefix.is_empty()).then(|| (k.to_string(), prefix.to_string()))
-        })
-        .collect();
-
-    let keys: Vec<String> = tools.iter().map(|(k, _)| k.to_string()).collect();
-    for k in keys {
-        tools.remove(&k);
-    }
-    for (k, v) in runtimes {
-        tools.insert(&k, v);
-    }
-    let first_linter_key = linters.first().map(|(k, _)| k.clone());
-    for (k, v) in linters {
-        tools.insert(&k, v);
-    }
-    for k in tools.iter().map(|(k, _)| k.to_string()).collect::<Vec<_>>() {
-        if let Some(mut key_mut) = tools.key_mut(&k) {
-            if let Some(prefix) = preserved_prefixes.get(&k) {
-                key_mut.leaf_decor_mut().set_prefix(prefix);
-                continue;
-            }
-            let existing_prefix = key_mut.leaf_decor().prefix().and_then(|prefix| {
-                prefix
-                    .as_str()
-                    .or_else(|| prefix.span().and_then(|span| original.get(span)))
-            });
-            if existing_prefix == Some("\n# Linters\n") {
-                key_mut.leaf_decor_mut().set_prefix("");
-            }
-        }
-    }
-    if let Some(k) = first_linter_key
-        && let Some(mut key_mut) = tools.key_mut(&k)
-    {
-        key_mut.leaf_decor_mut().set_prefix("\n# Linters\n");
-    }
-}
-
-const FLINT_V1_URL_PREFIX: &str = "https://raw.githubusercontent.com/grafana/flint/";
-
-pub(super) struct V1Removal {
-    /// Task keys that were removed from `[tasks]`.
-    pub removed_tasks: Vec<String>,
-    /// Whether `RENOVATE_TRACKED_DEPS_EXCLUDE` was removed from `[env]`.
-    pub removed_renovate_env: bool,
-    /// The manager list from `RENOVATE_TRACKED_DEPS_EXCLUDE`, split on commas, if it was present.
-    pub renovate_exclude_managers: Option<Vec<String>>,
-}
-
-/// Removes v1 HTTP task entries (tasks whose `file` value starts with the
-/// flint raw-GitHub URL) and, when a renovate-deps v1 task is present,
-/// also removes `RENOVATE_TRACKED_DEPS_EXCLUDE` from `[env]`.
-///
-/// Returns details about what was removed. Writes the file only when changed.
-pub(super) fn remove_v1_tasks(path: &Path) -> Result<V1Removal> {
-    let content = std::fs::read_to_string(path).unwrap_or_default();
-    let mut doc: toml_edit::DocumentMut = content
-        .parse()
-        .unwrap_or_else(|_| toml_edit::DocumentMut::new());
-
-    let mut removed_tasks: Vec<String> = Vec::new();
-    let mut has_v1_renovate = false;
-
-    if let Some(tasks) = doc.get_mut("tasks").and_then(|t| t.as_table_mut()) {
-        let keys_to_remove: Vec<String> = tasks
-            .iter()
-            .filter_map(|(key, item)| {
-                let file_val = item
-                    .as_table()
-                    .and_then(|t| t.get("file"))
-                    .and_then(|v| v.as_str())?;
-                if file_val.starts_with(FLINT_V1_URL_PREFIX) {
-                    Some(key.to_string())
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        for key in keys_to_remove {
-            // Check if it's a renovate-deps task before removing.
-            if let Some(file_val) = tasks
-                .get(&key)
-                .and_then(|i| i.as_table())
-                .and_then(|t| t.get("file"))
-                .and_then(|v| v.as_str())
-                && file_val.contains("renovate-deps")
-            {
-                has_v1_renovate = true;
-            }
-            tasks.remove(&key);
-            removed_tasks.push(key);
-        }
-    }
-
-    let mut removed_renovate_env = false;
-    let mut renovate_exclude_managers: Option<Vec<String>> = None;
-    if has_v1_renovate
-        && let Some(env) = doc.get_mut("env").and_then(|t| t.as_table_mut())
-        && let Some(val) = env
-            .get("RENOVATE_TRACKED_DEPS_EXCLUDE")
-            .and_then(|v| v.as_str())
-    {
-        renovate_exclude_managers = Some(
-            val.split(',')
-                .map(str::trim)
-                .filter(|s| !s.is_empty())
-                .map(str::to_string)
-                .collect(),
-        );
-        env.remove("RENOVATE_TRACKED_DEPS_EXCLUDE");
-        removed_renovate_env = true;
-    }
-
-    if !removed_tasks.is_empty() || removed_renovate_env {
-        std::fs::write(path, doc.to_string())?;
-    }
-
-    removed_tasks.sort();
-    Ok(V1Removal {
-        removed_tasks,
-        removed_renovate_env,
-        renovate_exclude_managers,
-    })
-}
+pub(super) use super::mise_tools::{
+    apply_changes, ensure_flint_self_pin, ensure_node_for_npm, remove_tool_keys,
+};
+pub(crate) use super::mise_tools::{
+    needs_node_for_npm, normalize_tools_section, replace_obsolete_keys,
+    tools_section_needs_normalization,
+};
+pub(super) use super::renovate::{flint_preset, patch_renovate_extends};
+pub(super) use super::v1::remove_v1_tasks;
 
 /// Returns true if any currently-selected check has `Category::Slow`.
 pub(super) fn has_slow_selected(groups: &[LinterGroup]) -> bool {
@@ -681,386 +87,5 @@ pub(super) fn prompt_config_dir(existing: Option<&str>, yes: bool) -> Result<Str
         Ok(path.trim().to_string())
     } else {
         Ok(CHOICES[idx.min(CHOICES.len() - 2)].to_string())
-    }
-}
-
-#[cfg(test)]
-mod node_prereq_tests {
-    use super::{ensure_flint_self_pin, ensure_node_for_npm, needs_node_for_npm};
-
-    #[test]
-    fn needs_node_when_npm_key_without_node() {
-        let content = "[tools]\n\"npm:renovate\" = \"43.129.0\"\n";
-        assert!(needs_node_for_npm(content));
-    }
-
-    #[test]
-    fn no_node_needed_when_no_npm_keys() {
-        let content = "[tools]\n\"github:koalaman/shellcheck\" = \"v0.11.0\"\n";
-        assert!(!needs_node_for_npm(content));
-    }
-
-    #[test]
-    fn no_node_needed_when_node_already_declared() {
-        let content = "[tools]\nnode = \"20\"\n\"npm:renovate\" = \"43.129.0\"\n";
-        assert!(!needs_node_for_npm(content));
-    }
-
-    #[test]
-    fn no_node_needed_when_tools_section_missing() {
-        assert!(!needs_node_for_npm(""));
-        assert!(!needs_node_for_npm("[env]\nFOO = \"bar\"\n"));
-    }
-
-    #[test]
-    fn noop_when_node_already_present() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        let original = "[tools]\nnode = \"20\"\n\"npm:renovate\" = \"43.129.0\"\n";
-        std::fs::write(&path, original).unwrap();
-        let added = ensure_node_for_npm(dir.path()).unwrap();
-        assert!(!added);
-        assert_eq!(std::fs::read_to_string(&path).unwrap(), original);
-    }
-
-    #[test]
-    fn noop_without_npm_tools() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        let original = "[tools]\n\"github:koalaman/shellcheck\" = \"v0.11.0\"\n";
-        std::fs::write(&path, original).unwrap();
-        let added = ensure_node_for_npm(dir.path()).unwrap();
-        assert!(!added);
-        assert_eq!(std::fs::read_to_string(&path).unwrap(), original);
-    }
-
-    #[test]
-    fn flint_self_pin_reverts_cargo_git_pin_to_release() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        std::fs::write(
-            &path,
-            "[tools]\n\"cargo:https://github.com/grafana/flint\" = \"rev:deadbeef\"\n\n[env]\nFLINT_CONFIG_DIR = \".github/config\"\n",
-        )
-        .unwrap();
-        let changed = ensure_flint_self_pin(dir.path(), None).unwrap();
-        let result = std::fs::read_to_string(&path).unwrap();
-        assert!(changed);
-        assert!(result.contains("\"github:grafana/flint\""));
-        assert!(!result.contains("cargo:https://github.com/grafana/flint"));
-        assert!(result.contains("FLINT_CONFIG_DIR = \".github/config\""));
-    }
-
-    #[test]
-    fn flint_self_pin_writes_cargo_git_pin_for_rev() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        std::fs::write(
-            &path,
-            "[tools]\n\"github:grafana/flint\" = \"0.20.5\"\n\n[env]\nFLINT_CONFIG_DIR = \".github/config\"\n",
-        )
-        .unwrap();
-        let changed = ensure_flint_self_pin(dir.path(), Some("rev:deadbeef")).unwrap();
-        let result = std::fs::read_to_string(&path).unwrap();
-        assert!(changed);
-        assert!(result.contains("\"cargo:https://github.com/grafana/flint\" = \"rev:deadbeef\""));
-        assert!(!result.contains("\"github:grafana/flint\""));
-        assert!(result.contains("FLINT_CONFIG_DIR = \".github/config\""));
-    }
-}
-
-#[cfg(test)]
-mod replace_obsolete_tests {
-    use super::{remove_tool_keys, replace_obsolete_keys};
-
-    #[test]
-    fn replaces_old_key_preserving_version() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        std::fs::write(&path, "[tools]\n\"github:mvdan/sh\" = \"v3.13.1\"\n").unwrap();
-        let replaced = replace_obsolete_keys(dir.path(), &[("github:mvdan/sh", "shfmt")]).unwrap();
-        assert_eq!(
-            replaced,
-            vec![("github:mvdan/sh".to_string(), "shfmt".to_string())]
-        );
-        let result = std::fs::read_to_string(&path).unwrap();
-        assert!(result.contains("shfmt"), "new key written: {result}");
-        assert!(
-            !result.contains("\"github:mvdan/sh\""),
-            "old key removed: {result}"
-        );
-        assert!(result.contains("v3.13.1"), "version preserved: {result}");
-    }
-
-    #[test]
-    fn noop_when_no_obsolete_keys() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        std::fs::write(&path, "[tools]\nshfmt = \"v3.13.1\"\n").unwrap();
-        let replaced = replace_obsolete_keys(dir.path(), &[("github:mvdan/sh", "shfmt")]).unwrap();
-        assert!(replaced.is_empty());
-    }
-
-    #[test]
-    fn replace_obsolete_keys_replaces_keys_without_normalizing_tools_section() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        std::fs::write(
-            &path,
-            "[tools]\nnode = \"24.15.0\"\n\n# Linters\n\"github:mvdan/sh\" = \"v3.13.1\"\n\"github:koalaman/shellcheck\" = \"v0.11.0\"\n",
-        )
-        .unwrap();
-        let replaced = replace_obsolete_keys(dir.path(), &[("github:mvdan/sh", "shfmt")]).unwrap();
-        assert_eq!(
-            replaced,
-            vec![("github:mvdan/sh".to_string(), "shfmt".to_string())]
-        );
-        let result = std::fs::read_to_string(&path).unwrap();
-        assert!(result.contains("node = \"24.15.0\""));
-        assert!(result.contains("\"github:koalaman/shellcheck\" = \"v0.11.0\""));
-        assert!(result.contains("shfmt = \"v3.13.1\""));
-    }
-
-    #[test]
-    fn removes_requested_tool_keys() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        std::fs::write(
-            &path,
-            "[tools]\n\"npm:prettier\" = \"3.6.2\"\n\"npm:markdownlint-cli2\" = \"0.18.1\"\n",
-        )
-        .unwrap();
-        let removed = remove_tool_keys(dir.path(), &["npm:prettier"]);
-        assert_eq!(removed.unwrap(), vec!["npm:prettier".to_string()]);
-        let result = std::fs::read_to_string(&path).unwrap();
-        assert!(!result.contains("npm:prettier"));
-        assert!(result.contains("npm:markdownlint-cli2"));
-    }
-
-    #[test]
-    fn remove_tool_keys_removes_entries_without_normalizing_tools_section() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("mise.toml");
-        std::fs::write(
-            &path,
-            "[tools]\nnode = \"24.15.0\"\n\n# Linters\n\"github:koalaman/shellcheck\" = \"v0.11.0\"\nshfmt = \"v3.13.1\"\n",
-        )
-        .unwrap();
-        let removed = remove_tool_keys(dir.path(), &["github:koalaman/shellcheck"]).unwrap();
-        assert_eq!(removed, vec!["github:koalaman/shellcheck".to_string()]);
-        let result = std::fs::read_to_string(&path).unwrap();
-        assert!(result.contains("node = \"24.15.0\""));
-        assert!(result.contains("shfmt = \"v3.13.1\""));
-        assert!(!result.contains("\"github:koalaman/shellcheck\" ="));
-    }
-}
-
-#[cfg(test)]
-mod v1_removal_tests {
-    use super::remove_v1_tasks;
-
-    fn write_tmp(content: &str) -> tempfile::NamedTempFile {
-        let f = tempfile::NamedTempFile::new().unwrap();
-        std::fs::write(f.path(), content).unwrap();
-        f
-    }
-
-    #[test]
-    fn removes_v1_http_tasks() {
-        let content = r#"
-[tools]
-lychee = "latest"
-
-[tasks."lint:links"]
-description = "Check for broken links"
-file = "https://raw.githubusercontent.com/grafana/flint/abc123/tasks/lint/links.sh"
-
-[tasks."lint:renovate-deps"]
-description = "Check renovate deps"
-file = "https://raw.githubusercontent.com/grafana/flint/abc123/tasks/lint/renovate-deps.py"
-
-[tasks.build]
-description = "Build the project"
-run = "cargo build"
-"#;
-        let tmp = write_tmp(content);
-        let result = remove_v1_tasks(tmp.path()).unwrap();
-        assert_eq!(result.removed_tasks, ["lint:links", "lint:renovate-deps"]);
-        assert!(!result.removed_renovate_env);
-        let after = std::fs::read_to_string(tmp.path()).unwrap();
-        assert!(!after.contains("lint:links"));
-        assert!(!after.contains("lint:renovate-deps"));
-        assert!(after.contains("[tasks.build]"), "non-v1 tasks preserved");
-    }
-
-    #[test]
-    fn removes_renovate_env_when_v1_renovate_task_present() {
-        let content = r#"
-[env]
-RENOVATE_TRACKED_DEPS_EXCLUDE = "github-actions, github-runners"
-
-[tasks."lint:renovate-deps"]
-description = "Check renovate deps"
-file = "https://raw.githubusercontent.com/grafana/flint/abc123/tasks/lint/renovate-deps.py"
-"#;
-        let tmp = write_tmp(content);
-        let result = remove_v1_tasks(tmp.path()).unwrap();
-        assert_eq!(result.removed_tasks, ["lint:renovate-deps"]);
-        assert!(result.removed_renovate_env);
-        assert_eq!(
-            result.renovate_exclude_managers,
-            Some(vec![
-                "github-actions".to_string(),
-                "github-runners".to_string()
-            ])
-        );
-        let after = std::fs::read_to_string(tmp.path()).unwrap();
-        assert!(!after.contains("RENOVATE_TRACKED_DEPS_EXCLUDE"));
-    }
-
-    #[test]
-    fn does_not_remove_renovate_env_without_v1_renovate_task() {
-        let content = r#"
-[env]
-RENOVATE_TRACKED_DEPS_EXCLUDE = "github-actions"
-
-[tasks."lint:links"]
-description = "Check links"
-file = "https://raw.githubusercontent.com/grafana/flint/abc123/tasks/lint/links.sh"
-"#;
-        let tmp = write_tmp(content);
-        let result = remove_v1_tasks(tmp.path()).unwrap();
-        assert_eq!(result.removed_tasks, ["lint:links"]);
-        assert!(!result.removed_renovate_env);
-        let after = std::fs::read_to_string(tmp.path()).unwrap();
-        assert!(
-            after.contains("RENOVATE_TRACKED_DEPS_EXCLUDE"),
-            "env var preserved when no renovate task"
-        );
-    }
-
-    #[test]
-    fn no_op_when_no_v1_tasks() {
-        let content = "[tools]\nlychee = \"latest\"\n";
-        let tmp = write_tmp(content);
-        let original_mtime = std::fs::metadata(tmp.path()).unwrap().modified().unwrap();
-        let result = remove_v1_tasks(tmp.path()).unwrap();
-        assert!(result.removed_tasks.is_empty());
-        assert!(!result.removed_renovate_env);
-        // File should not have been written.
-        let new_mtime = std::fs::metadata(tmp.path()).unwrap().modified().unwrap();
-        assert_eq!(
-            original_mtime, new_mtime,
-            "file unchanged when nothing to remove"
-        );
-    }
-
-    #[test]
-    fn ignores_non_flint_http_tasks() {
-        let content = r#"
-[tasks."lint:something"]
-file = "https://raw.githubusercontent.com/some-other-org/some-repo/abc123/task.sh"
-"#;
-        let tmp = write_tmp(content);
-        let result = remove_v1_tasks(tmp.path()).unwrap();
-        assert!(result.removed_tasks.is_empty());
-    }
-}
-
-#[cfg(test)]
-mod extends_tests {
-    use super::{add_to_extends, patch_renovate_extends};
-
-    fn write_tmp(content: &str) -> tempfile::NamedTempFile {
-        let f = tempfile::NamedTempFile::new().unwrap();
-        std::fs::write(f.path(), content).unwrap();
-        f
-    }
-
-    #[test]
-    fn replaces_unpinned_flint_entry_in_place() {
-        let input = r#"{ extends: ["config:recommended", "github>grafana/flint"] }"#;
-        let tmp = write_tmp(input);
-        let changed = patch_renovate_extends(tmp.path()).unwrap();
-        assert!(changed);
-        let result = std::fs::read_to_string(tmp.path()).unwrap();
-        assert!(
-            result.contains("github>grafana/flint#v"),
-            "pinned entry written: {result}"
-        );
-        // Only one flint entry — no duplicate
-        assert_eq!(
-            result.matches("grafana/flint").count(),
-            1,
-            "no duplicate: {result}"
-        );
-        assert!(
-            !result.contains("\"github>grafana/flint\""),
-            "unpinned removed: {result}"
-        );
-    }
-
-    #[test]
-    fn replaces_differently_pinned_flint_entry() {
-        let input = r#"{ extends: ["config:recommended", "github>grafana/flint#v0.5.0"] }"#;
-        let tmp = write_tmp(input);
-        let changed = patch_renovate_extends(tmp.path()).unwrap();
-        assert!(changed);
-        let result = std::fs::read_to_string(tmp.path()).unwrap();
-        assert!(!result.contains("v0.5.0"), "old pin removed: {result}");
-        assert_eq!(
-            result.matches("grafana/flint").count(),
-            1,
-            "no duplicate: {result}"
-        );
-    }
-
-    #[test]
-    fn no_op_when_already_pinned_to_current_version() {
-        let entry = super::flint_preset();
-        let input = format!(r#"{{ extends: ["config:recommended", "{entry}"] }}"#);
-        let tmp = write_tmp(&input);
-        let changed = patch_renovate_extends(tmp.path()).unwrap();
-        assert!(!changed);
-    }
-
-    #[test]
-    fn adds_to_single_line_extends() {
-        let input = r#"{ "extends": ["config:recommended"], "other": 1 }"#;
-        let result = add_to_extends(input, "github>grafana/flint#v0.9.2").unwrap();
-        assert!(result.contains(r#"["github>grafana/flint#v0.9.2", "config:recommended"]"#));
-    }
-
-    #[test]
-    fn adds_to_json5_unquoted_key() {
-        let input = "{\n  extends: [\"config:recommended\"],\n}\n";
-        let result = add_to_extends(input, "github>grafana/flint#v0.9.2").unwrap();
-        assert!(result.contains(r#""github>grafana/flint#v0.9.2", "config:recommended""#));
-    }
-
-    #[test]
-    fn adds_to_multiline_extends() {
-        let input = "{\n  extends: [\n    \"config:recommended\",\n    \"other\"\n  ]\n}\n";
-        let result = add_to_extends(input, "github>grafana/flint#v0.9.2").unwrap();
-        assert!(result.contains("\"github>grafana/flint#v0.9.2\","));
-        // Entry should appear before existing entries
-        let flint_pos = result.find("grafana/flint").unwrap();
-        let existing_pos = result.find("config:recommended").unwrap();
-        assert!(flint_pos < existing_pos);
-    }
-
-    #[test]
-    fn adds_extends_when_absent() {
-        let input = "{\n  \"branchPrefix\": \"renovate/\"\n}\n";
-        let result = add_to_extends(input, "github>grafana/flint#v0.9.2").unwrap();
-        assert!(result.contains("\"extends\""));
-        assert!(result.contains("github>grafana/flint#v0.9.2"));
-    }
-
-    #[test]
-    fn adds_to_empty_extends_array() {
-        let input = r#"{ "extends": [] }"#;
-        let result = add_to_extends(input, "github>grafana/flint#v0.9.2").unwrap();
-        assert!(result.contains(r#"["github>grafana/flint#v0.9.2"]"#));
     }
 }
